tests: add option to run tests without docker
via `export OPENSSH_NO_DOCKER=1`.

SSH server host can be set via:
  `export OPENSSH_SERVER_HOST=127.0.0.1`

SSH server port via existing:
  `export OPENSSH_SERVER_PORT=4711`

This requires more work to be usable out of the box. The necessery sshd
config is (partly) embedded into `tests/openssh_server/Dockerfile`.

After this patch, it is possible to run tests in envs where docker is
not installed or not available, by running a preconfigured,
non-containerized sshd.

+++ tests/openssh_fixture.c
@@ -59,6 +59,8 @@
 #include <string.h>
 #include <stdarg.h>
 
+static int have_docker = 0;
+
 static int run_command_varg(char **output, const char *command, va_list args)
 {
     FILE *pipe;
@@ -150,22 +152,27 @@ static const char *openssh_server_image(void)
 
 static int build_openssh_server_docker_image(void)
 {
-    const char *container_image_name = openssh_server_image();
-    if(container_image_name != NULL) {
-        int ret = run_command(NULL, "docker pull --quiet %s",
-                              container_image_name);
-        if(ret == 0) {
-            ret = run_command(NULL, "docker tag %s libssh2/openssh_server",
-                              container_image_name);
+    if(have_docker) {
+        const char *container_image_name = openssh_server_image();
+        if(container_image_name != NULL) {
+            int ret = run_command(NULL, "docker pull --quiet %s",
+                                  container_image_name);
             if(ret == 0) {
-                return ret;
+                ret = run_command(NULL, "docker tag %s libssh2/openssh_server",
+                                  container_image_name);
+                if(ret == 0) {
+                    return ret;
+                }
             }
         }
-    }
 
-    return run_command(NULL, "docker build --quiet "
-                             "-t libssh2/openssh_server "
-                             "openssh_server");
+        return run_command(NULL, "docker build --quiet "
+                                 "-t libssh2/openssh_server "
+                                 "openssh_server");
+    }
+    else {
+        return 0;
+    }
 }
 
 static const char *openssh_server_port(void)
@@ -175,22 +182,33 @@ static const char *openssh_server_port(void)
 
 static int start_openssh_server(char **container_id_out)
 {
-    const char *container_host_port = openssh_server_port();
-    if(container_host_port != NULL) {
-        return run_command(container_id_out,
-                           "docker run --rm -d -p %s:22 "
-                           "libssh2/openssh_server",
-                           container_host_port);
-    }
+    if(have_docker) {
+        const char *container_host_port = openssh_server_port();
+        if(container_host_port != NULL) {
+            return run_command(container_id_out,
+                               "docker run --rm -d -p %s:22 "
+                               "libssh2/openssh_server",
+                               container_host_port);
+        }
 
-    return run_command(container_id_out,
-                       "docker run --rm -d -p 22 "
-                       "libssh2/openssh_server");
+        return run_command(container_id_out,
+                           "docker run --rm -d -p 22 "
+                           "libssh2/openssh_server");
+    }
+    else {
+        *container_id_out = strdup("");
+        return 0;
+    }
 }
 
 static int stop_openssh_server(char *container_id)
 {
-    return run_command(NULL, "docker stop %s", container_id);
+    if(have_docker) {
+        return run_command(NULL, "docker stop %s", container_id);
+    }
+    else {
+        return 0;
+    }
 }
 
 static const char *docker_machine_name(void)
@@ -307,20 +325,37 @@ static int open_socket_to_container(char *container_id)
     int sock;
     struct sockaddr_in sin;
     int counter = 0;
+    int ret;
 
-    int ret = ip_address_from_container(container_id, &ip_address);
-    if(ret != 0) {
-        fprintf(stderr, "Failed to get IP address for container %s\n",
-                container_id);
-        ret = -1;
-        goto cleanup;
+    if(have_docker) {
+        ret = ip_address_from_container(container_id, &ip_address);
+        if(ret != 0) {
+            fprintf(stderr, "Failed to get IP address for container %s\n",
+                    container_id);
+            ret = -1;
+            goto cleanup;
+        }
+
+        ret = port_from_container(container_id, &port_string);
+        if(ret != 0) {
+            fprintf(stderr, "Failed to get port for container %s\n",
+                    container_id);
+            ret = -1;
+        }
     }
-
-    ret = port_from_container(container_id, &port_string);
-    if(ret != 0) {
-        fprintf(stderr, "Failed to get port for container %s\n",
-                container_id);
-        ret = -1;
+    else {
+        const char *env;
+        env = getenv("OPENSSH_SERVER_HOST");
+        if(!env) {
+            env = "127.0.0.1";
+        }
+        ip_address = strdup(env);
+        env = openssh_server_port();
+        if(!env) {
+            env = "4711";
+        }
+        port_string = strdup(env);
+        ret = 0;
     }
 
     /* 0.0.0.0 is returned by Docker for Windows, because the container
@@ -391,6 +426,8 @@ int start_openssh_fixture()
     }
 #endif
 
+    have_docker = (getenv("OPENSSH_NO_DOCKER") == NULL);
+
     ret = build_openssh_server_docker_image();
     if(ret == 0) {
         return start_openssh_server(&running_container_id);
@@ -408,7 +445,7 @@ void stop_openssh_fixture()
         free(running_container_id);
         running_container_id = NULL;
     }
-    else {
+    else if(have_docker) {
         fprintf(stderr, "Cannot stop container - none started");
     }
 }
